requirements.txt
faster-whisper
moviepy
google-generativeai
python-docx
pypdf
python-dotenv
mutagen
